move event and batch data classes to a model group
