+++ ios/sdk/O2MC/Models/O2MEvent.m
