+++ ios/sdk/O2MC/O2MEventManager.h
@@ -7,7 +7,7 @@
 //
 
 #import <Foundation/Foundation.h>
-#import "O2MEvent.h"
+#import "Models/O2MEvent.h"
 
 @interface O2MEventManager : NSObject
